% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/directorio.R
\name{actualizarDirectorio}
\alias{actualizarDirectorio}
\title{Actualizar el Directorio despues de generar una tabla}
\usage{
actualizarDirectorio(DB, Directorio = Directorio)
}
\arguments{
\item{DB}{Tabla de datos a introducir al Directorio}

\item{Directorio}{Tabla de Directorio, guardada en una Variable global}
}
\value{
Modifica al objeto Directorio
}
\description{
Ver 2.4 - Genera una copia fisica del directorio.
Ver 2.3 - Función que no necesita ser reasignada a una variable
}
