% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/directorio.R
\name{describirDirectorio}
\alias{describirDirectorio}
\title{Cuenta los elementos del Directorio segun tema y tipo}
\usage{
describirDirectorio()
}
\value{
Genera, en mensajes, un listado de los elementos contenidos en el Directorio
}
\description{
Cuenta los elementos del Directorio segun tema y tipo
}
