% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interfaz.R
\name{doeExam}
\alias{doeExam}
\title{Interfaz de usuario: Generador de Examenes de Diseño de Experimentos}
\usage{
doeExam()
}
\description{
Ver 1.1.0 - Correccion para los casos donde el directorio existe, pero esta vacio
Ver 1.0.2 -
}
