% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doe.testR-package.R
\docType{package}
\name{DOEtestR-package}
\alias{DOEtestR}
\alias{DOEtestR-package}
\title{DOEtestR: Generador de examenes de Diseño de Experimentos}
\description{
Contiene funciones para generar ejercicios y examenes para un curso de Diseño de Experimentos. En la versión actual, contiene las funcionalidades para desarrollar ejercicios de Media Muestral, Prueba de Hipotesis para comparación de medias, Anova de 1 Factor y Pruebas Post-Hoc. Incluye desarrollo de un sistema de directorio de tablas relacionales para el manejo de todos los ejercicios y examenes desarrollados. Para la version 2.0.0 se diseñó un Command Line Interface para facilitar el uso de la libreria - para usarla, se llama a la variable ´doeExam()´.
}
\author{
\strong{Maintainer}: Francisco Castillo-Escalante \email{castillopaco@gmail.com} (\href{https://orcid.org/0000-0003-4267-3281}{ORCID})

}
\keyword{internal}
