% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/database.R
\name{exportDB}
\alias{exportDB}
\title{Exportar tabla de datos}
\usage{
exportDB(DB)
}
\arguments{
\item{DB}{Tabla de datos a exportar}
}
\description{
V2.1 - usar los paquetes here y folders para los paths relativos
V2 - usar saveRDS y readRDS para los archivos
V1 - usar dput y dget para los archivos
}
