% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generadores.R
\name{gen1Anova}
\alias{gen1Anova}
\title{Anova de 1 Factor: Generador de datos muestrales}
\usage{
gen1Anova(r)
}
\arguments{
\item{r}{Numero de sets de datos a generar}
}
\value{
Una tabla de datos, se guarda en un archivo, y se puede asignar a un objeto
}
\description{
Version 3.6.1
}
