% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generadores.R
\name{gen1FBlock}
\alias{gen1FBlock}
\title{Anova 1F + Bloque: Generador de datos muestrales}
\usage{
gen1FBlock(r)
}
\arguments{
\item{r}{Numero de sets de datos a generar}
}
\value{
Una tabla de datos, se guarda en un archivo, y se puede asignar a un objeto
}
\description{
Ver 2.1.1 - Cambios menores para importar funciones externas
Ver 2.1 - Incluye integracion al directorio.
}
