% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generadores.R
\name{gen2k}
\alias{gen2k}
\title{Diseño 2^k: Generador de datos muestrales}
\usage{
gen2k(r)
}
\arguments{
\item{r}{Numero de sets de datos a generar}
}
\value{
Una tabla de datos, se guarda en un archivo, y se puede asignar a un objeto
}
\description{
Ver 2.1.0 - Integración del directorio
Ver 2.0.0 - Inicial
}
