% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generadores.R
\name{genMediaMuestral}
\alias{genMediaMuestral}
\title{Media Muestral: Generador de datos muestrales}
\usage{
genMediaMuestral(r)
}
\arguments{
\item{r}{Numero de sets de datos a generar}
}
\value{
Una tabla de datos, se guarda en un archivo, y se puede asignar a un objeto
}
\description{
Version 3.7.3 - usar los paquetes here y folders para los paths relativos.
Version 3.7.2
}
