% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/examen.R
\name{genPractica}
\alias{genPractica}
\title{Generador de Ejercicios de practica por tema}
\usage{
genPractica(tema = NULL, format = "HTML", lang = "Esp")
}
\arguments{
\item{tema}{Segun el tema: 1 (Media muestral), 2 (Pruebas de hipotesis), 3 (Anova 1F), 4 (Anova 1F+B), 5 (Diseño 2^k)}

\item{format}{HTML o LaTeX, formato en que se generara el archivo de salida del examen}

\item{lang}{Esp o Eng. Idioma en que se armara el archivo de salida}
}
\value{
Genera una tabla con los enunciados del ejercicio, un archivo de salida de texto (HTML o LaTeX), y actualiza el directorio
}
\description{
Ver 0.2.0 - Se incorporan el framework para trabajar con los temas B y K
Version 0.1.0 - Version inicial
}
