% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generadores.R
\name{genPruebaHipotesis}
\alias{genPruebaHipotesis}
\title{Prueba de Hipotesis: Generador de datos muestrales}
\usage{
genPruebaHipotesis(r)
}
\arguments{
\item{r}{Numero de sets de datos a generar}
}
\value{
Una tabla de datos, se guarda en un archivo, y se puede asignar a un objeto
}
\description{
Version 3.7.2
}
