% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/database.R
\name{parsearQuery}
\alias{parsearQuery}
\title{Parsear query de id de la tabla de datos}
\usage{
parsearQuery(query)
}
\arguments{
\item{query}{String, puede ser el nombre de la tabla completo, incompleto. Si se presentan mas de uno, se genera un menu}
}
\value{
Otorga el nombre oficial de la tabla DBname, con el formato \link{Tema}-\link{Tipo}-\link{Fecha}-\link{id_byte}-\link{Idioma, si se requiere}
}
\description{
Version 2.2 - Mejorado para los casos donde se presentan más de un match, no se presentan matches, y se da un query incompleto.
}
