% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generadores.R
\name{question1Anova}
\alias{question1Anova}
\title{Anova de 1 Factor: Generador de preguntas}
\usage{
question1Anova(Data, Q, format = "HTML", lang = "Esp", silent = T)
}
\arguments{
\item{Data}{Tabla de datos muestrales a usar para la generación de preguntas}

\item{Q}{Numero de pregunas a generar}

\item{format}{Formato para los enunciados de preguntas. "HTML" o "LaTeX".}

\item{lang}{Eng o Esp. Idiomas para los enunciados}

\item{silent}{Si \code{TRUE}, enonces no se abre el archivo de texto generado}
}
\value{
Una tabla de datos, se guarda en un archivo, y se puede asignar a un objeto
}
\description{
Version 3.8.5 - Se cambian los ´on.exit(sink())´ y ´sink()´ por  ´closeAllConnections()´
Version 3.8.4 - usar los paquetes here y folders para los paths relativos
Version 3.8.3 - Adaptado para trabajar en un package. Tabla "statement" ahora se guarda dentro del objeto interno "sysdata.rda"
Version 3.8.2 - Adaptado para generar enunciados en ingles y español simultaneamente
}
