% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generadores.R
\name{questionMediaMuestral}
\alias{questionMediaMuestral}
\title{Media Muestral: Generador de preguntas}
\usage{
questionMediaMuestral(Data, Q, format = "HTML", silent = T)
}
\arguments{
\item{Data}{Tabla de datos muestrales a usar para la generación de preguntas}

\item{Q}{Numero de pregunas a generar}

\item{format}{Formato para los enunciados de preguntas. "HTML" o "LaTeX".}

\item{silent}{Si \code{TRUE}, enonces no se abre el archivo de texto generado}
}
\value{
Una tabla de datos, se guarda en un archivo, y se puede asignar a un objeto
}
\description{
Vers 3.7.3 - Ajuste del foermato del enunciado  para incorporar  HTML Entities
Vers 3.7.2 - Se cambian los ´on.exit(sink())´ y ´sink()´ por  ´closeAllConnections()´
Version 3.7.1 - Adaptado para generar enunciados en ingles y español simultaneamente
}
