% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generadores.R
\name{questionPruebaHipotesis}
\alias{questionPruebaHipotesis}
\title{Prueba de Hipotesis: Generador de preguntas}
\usage{
questionPruebaHipotesis(Data, Q, format = "HTML", silent = T)
}
\arguments{
\item{Data}{Tabla de datos muestrales a usar para la generación de preguntas}

\item{Q}{Numero de pregunas a generar}

\item{format}{Formato para los enunciados de preguntas. "HTML" o "LaTeX".}

\item{silent}{Si \code{TRUE}, enonces no se abre el archivo de texto generado}
}
\value{
Una tabla de datos, se guarda en un archivo, y se puede asignar a un objeto
}
\description{
Ver 3.6.6 - Ajuste del foermato del enunciado  para incorporar  HTML Entities
Ver 3.6.5 - Se cambian los ´on.exit(sink())´ y ´sink()´ por  ´closeAllConnections()´
Ver 3.6.4 - usar los paquetes here y folders para los paths relativos
Ver 3.6.3 - Se trata de corregir el problema con las pruebas de dos colas
Version 3.6.1 - Adaptado para generar enunciados en ingles y español simultaneamente
}
