% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reportes.R
\name{rep_D_2k}
\alias{rep_D_2k}
\title{Diseño 2^k: Reporte de sets de datos}
\usage{
rep_D_2k(DB, out)
}
\arguments{
\item{DB}{Tabla de datos muestrales}
}
\value{
Genera un archivo de texto
}
\description{
Diseño 2^k: Reporte de sets de datos
}
