% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reportes.R
\name{rep_D_A1}
\alias{rep_D_A1}
\title{Anova de 1 Factor: Reporte de la tabla de datos muestrales}
\usage{
rep_D_A1(DB)
}
\arguments{
\item{DB}{tabla de datos del tema Anova de 1 Factor}
}
\value{
Genera un archivo de texto
}
\description{
Anova de 1 Factor: Reporte de la tabla de datos muestrales
}
