% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reportes.R
\name{rep_D_AB}
\alias{rep_D_AB}
\title{Anova 1F + Bloque: Reporte de sets de datos}
\usage{
rep_D_AB(DB)
}
\arguments{
\item{DB}{Tabla de datos muestrales}
}
\value{
Genera un archivo de texto
}
\description{
Anova 1F + Bloque: Reporte de sets de datos
}
