% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reportes.R
\name{rep_D_MM}
\alias{rep_D_MM}
\title{Media Muestral: Reporte de la tabla de datos muestrales}
\usage{
rep_D_MM(DB)
}
\arguments{
\item{DB}{tabla de datos del tema Media Muestral}
}
\value{
Genera un archivo de texto
}
\description{
Media Muestral: Reporte de la tabla de datos muestrales
}
