% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reportes.R
\name{rep_D_PH}
\alias{rep_D_PH}
\title{Prueba de Hipotesis: Reporte de la tabla de datos muestrales}
\usage{
rep_D_PH(DB)
}
\arguments{
\item{DB}{tabla de datos del tema Prueba de Hipotesis}
}
\value{
Genera un archivo de texto
}
\description{
Prueba de Hipotesis: Reporte de la tabla de datos muestrales
}
