% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/examen.R
\name{rep_Exam}
\alias{rep_Exam}
\title{Reporte de respuestas del examen}
\usage{
rep_Exam(Data = NULL)
}
\arguments{
\item{Data}{Pide una tabla de enunciados de examen, en caso de no darla, pide un id para buscar en el directorio.}
}
\value{
Genera un archivo de texto con las preguntas del examen.
}
\description{
Ver 2.3.1  - Se cambian los ´on.exit(sink())´ y ´sink()´ por  ´closeAllConnections()´
Ver 2.3 - usar los paquetes here y folders para los paths relativos
Ver 2.2 - Se altera el input, de tal forma que se pueda usar la función justo despues de crear un examen.
Ver 2.1
}
