% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reportes.R
\name{rep_PostHoc}
\alias{rep_PostHoc}
\title{Anova de 1 Factor: Reporte de resultados de Pruebas Post Hoc}
\usage{
rep_PostHoc(Data)
}
\arguments{
\item{Data}{Fila de la tabla de datos a describir}
}
\description{
Anova de 1 Factor: Reporte de resultados de Pruebas Post Hoc
}
