% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reportes.R
\name{rep_Q_A1}
\alias{rep_Q_A1}
\title{Anova de 1 Factor: Reporte de preguntas generadas}
\usage{
rep_Q_A1(Data, Ques, i)
}
\arguments{
\item{Data}{Tabla de datos muestrales}

\item{Ques}{Tabla de ejercicios generados a parir de la tabla de datos muestrales}

\item{i}{Id de la pregunta especifica a describir}
}
\description{
Anova de 1 Factor: Reporte de preguntas generadas
}
