% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reportes.R
\name{rep_Q_AB}
\alias{rep_Q_AB}
\title{Anova 1F + Bloque: Reporte de preguntas elaboradas}
\usage{
rep_Q_AB(Data, Ques, i)
}
\arguments{
\item{Data}{Tabla de datos muestrales}

\item{Ques}{Tabla de ejercicios generados a parir de la tabla de datos muestrales}

\item{i}{Id de la pregunta especifica a describir}
}
\value{
Genera un archivo de texto
}
\description{
Anova 1F + Bloque: Reporte de preguntas elaboradas
}
