% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reportes.R
\name{rep_Q_PH}
\alias{rep_Q_PH}
\title{Prueba de Hipotesis: Reporte de preguntas generadas}
\usage{
rep_Q_PH(Data, Ques, i)
}
\arguments{
\item{Data}{Tabla de datos muestrales}

\item{Ques}{Tabla de ejercicios generados a parir de la tabla de datos muestrales}

\item{i}{Id de la pregunta especifica a describir}
}
\description{
Prueba de Hipotesis: Reporte de preguntas generadas
}
