% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/directorio.R
\name{resetDirectorio}
\alias{resetDirectorio}
\title{Reinicia el Directorio con las tablas guardadas en la memoria}
\usage{
resetDirectorio()
}
\value{
Vuelve a genera la variable Directorio a partir de las tablas de datos existentes
}
\description{
Vers 3.0 Se ajusta para los casos de inicialización del directorio. (i.e. El directorio esta vacio).
Vers 2.0 Se cambia para usar el los paquetes here y folders
}
