% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interfaz.R
\name{poblarDirectorio}
\alias{poblarDirectorio}
\title{Crear tablas para un Directorio.}
\usage{
poblarDirectorio(parcial = NULL, aislado = FALSE)
}
\arguments{
\item{parcial}{Numero del parcial a genera (1 o 2)}

\item{aislado}{Flag para saber si se genera un Directorio temporal o no.}
}
\description{
Ver 1.0.0
}
